+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## 3.5.1
 * `CHG` setting `type.castNumberToInteger` default by `true`
+* `CHG` improve supports for multi-workspace
 * `FIX` [#1335](https://github.com/sumneko/lua-language-server/issues/1335)
 * `FIX` [#1354](https://github.com/sumneko/lua-language-server/issues/1354)